backends/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
eventlet==0.40.3
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-PyMongo==3.0.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
geojson==3.2.0
googlemaps==4.10.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
openai==1.107.2
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.15.0
pyowm==3.5.0
PySocks==1.7.1
python-dotenv==1.1.1
python-engineio==4.12.2
python-http-client==3.3.7
python-socketio==5.13.0
requests==2.32.5
sendgrid==6.12.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
